Fix link comparison
Links could previously change position, and the values of points 1 and 2, leading to the possibility of links that display in the same place but with different values, and failure to detect similarity.

Now links always have one point at (0, 0), and use the position of the link to record where to shift it to. The link will always point right or down from its position.